+++ app/swipe_session/services/swipe_session_websocket.py
@@ -6,6 +6,7 @@ from fastapi import WebSocket, WebSocketDisconnect, WebSocketException
 from pydantic import ValidationError
 from starlette.websockets import WebSocketState
 from app.group.services.group import GroupService
+from app.recipe.exceptions.recipe import RecipeNotFoundException
 from app.recipe.schemas.recipe import GetFullRecipeResponseSchema
 from app.recipe.services.recipe import RecipeService
 from app.swipe.schemas.swipe import CreateSwipeSchema
@@ -27,7 +28,6 @@ from core.exceptions.websocket import (
 from core.db.enums import SwipeSessionActionEnum, SwipeSessionEnum
 from core.db.models import SwipeSession, User
 from core.exceptions.base import CustomException, UnauthorizedException
-from core.exceptions.recipe import RecipeNotFoundException
 from core.helpers.hashid import decode_single
 from core.helpers.websocket.auth import (
     IsActiveSession,
