+++ core/exceptions/__init__.py
@@ -7,40 +7,21 @@ from .base import (
     DuplicateValueException,
     UnauthorizedException,
 )
-from .group import GroupNotFoundException, GroupJoinConflictException
 from .token import DecodeTokenException, ExpiredTokenException
-from app.user.exceptions.user import (
-    PasswordDoesNotMatchException,
-    DuplicateUsernameException,
-    UserNotFoundException,
-    MissingUserIDException,
-    MissingGroupIDException,
-    IncorrectPasswordException,
-)
-from .recipe import RecipeNotFoundException
 from .responses import ExceptionResponseSchema
 from .hashids import IncorrectHashIDException
 
 
 __all__ = [
-    "BadRequestException",
     "CustomException",
-    "DecodeTokenException",
-    "UnprocessableEntity",
-    "DuplicateUsernameException",
-    "DuplicateValueException",
-    "ExceptionResponseSchema",
-    "ExpiredTokenException",
-    "ForbiddenException",
-    "GroupNotFoundException",
-    "GroupJoinConflictException",
-    "IncorrectHashIDException",
-    "IncorrectPasswordException",
-    "MissingGroupIDException",
-    "MissingUserIDException",
+    "BadRequestException",
     "NotFoundException",
-    "PasswordDoesNotMatchException",
-    "RecipeNotFoundException",
+    "ForbiddenException",
+    "UnprocessableEntity",
+    "DuplicateValueException",
     "UnauthorizedException",
-    "UserNotFoundException",
+    "DecodeTokenException",
+    "ExpiredTokenException",
+    "ExceptionResponseSchema",
+    "IncorrectHashIDException",
 ]
